+++ imml/decomposition/mofa.py
@@ -175,7 +175,7 @@ class MOFA(TransformerMixin, BaseEstimator):
         Xs : list of array-likes objects
             - Xs length: n_mods
             - Xs[i] shape: (n_samples_i, n_features_i)
-            A list of different mods.
+            A list of different data modalities.
 
         Returns
         -------
@@ -204,7 +204,7 @@ class MOFA(TransformerMixin, BaseEstimator):
         Xs : list of array-likes objects
             - Xs length: n_mods
             - Xs[i] shape: (n_samples_i, n_features_i)
-            A list of different mods.
+            A list of different data modalities.
         y : Ignored
             Not used, present here for API consistency by convention.
         fit_params : Ignored