+++ tests/preprocessing/test_select_incomplete_samples.py
@@ -31,8 +31,8 @@ def test_select_incomplete_samples_class(sample_data):
         transformed_Xs = transformer.fit_transform(Xs)
         type(transformed_Xs[0]) is type(Xs[0])
         expected_values = [4, 4, 4]
-        for transformed, expected_value in zip(transformed_Xs, expected_values):
-            np.equal(transformed, expected_value)
+        for transformed_X, expected_value in zip(transformed_Xs, expected_values):
+            assert len(transformed_X) == expected_value
 
 
 def test_select_incomplete_samples_function(sample_data):
@@ -40,8 +40,24 @@ def test_select_incomplete_samples_function(sample_data):
         transformed_Xs = select_incomplete_samples(Xs)
         type(transformed_Xs[0]) is type(Xs[0])
         expected_values = [4, 4, 4]
-        for transformed, expected_value in zip(transformed_Xs, expected_values):
-            np.equal(transformed, expected_value)
+        for transformed_X, expected_value in zip(transformed_Xs, expected_values):
+            assert len(transformed_X) == expected_value
+
+
+def test_select_incomplete_samples_return_y(sample_data):
+    for Xs in sample_data:
+        y = np.random.default_rng(42).integers(0, 2, len(Xs[0]))
+        if isinstance(Xs[0], pd.DataFrame):
+            y = pd.Series(y, index=Xs[0].index)
+        elif deepmodule_installed:
+            if isinstance(Xs[0], torch.Tensor):
+                y = torch.from_numpy(y)
+        transformed_Xs, y = select_incomplete_samples(Xs, y=y)
+        type(transformed_Xs[0]) is type(Xs[0])
+        expected_values = [4, 4, 4]
+        for transformed_X, expected_value in zip(transformed_Xs, expected_values):
+            assert len(transformed_X) == expected_value
+            assert len(y) == expected_value
 
 
 def test_select_incomplete_samples_multiple_types(sample_data):
@@ -60,7 +76,7 @@ def test_select_incomplete_samples_multiple_types(sample_data):
         type(transformed_Xs[0]) is type(Xs[0])
         expected_values = [4, 4, 4]
         for transformed_X, expected_value in zip(transformed_Xs, expected_values):
-            np.equal(transformed_X, expected_value)
+            assert len(transformed_X) == expected_value
 
 
 if __name__ == "__main__":